refactor: ScheduleValidation -> ScheduleCreateValidation
