--- src/main/java/nextstep/application/schedule/ScheduleValidation.java
@@ -1,10 +1,0 @@
-package nextstep.application.schedule;
-
-import nextstep.application.schedule.dto.Schedule;
-
-public interface ScheduleValidation {
-
-  void checkValid(Schedule schedule);
-
-  int priority();
-}
